Support partial updates to movies
This commit allows clients to send partial updates to movies. The
`PUT /v1/movies/:id` endpoint now accepts a JSON request body that
contains a subset of the fields in the movie struct. The handler
merges the existing movie with the new values, and then validates
the result.

The endpoint is now hit with a `PATCH` request instead of a `PUT`
method. This is a more accurate description of what the endpoint
does.