Gracefully shut down background goroutines
This commit adds a wg dependency to the app. The WaitGroup is incremented
each time before a background task is run. And, when the goroutine responsible
for intercepting SIGTERM and SIGINT signals finally does so, it waits for
the WaitGroup to be zero before passing nil to the shutdownErr channel.