Create authentication token
This commit adds a new endpoint to the API that allows a user to
authenticate and receive a token that can be used to access other
endpoints.

The endpoint is /v1/tokens/authentication and accepts a
POST request with a JSON request body containing the user's email
address and password. If the credentials are valid, the endpoint
returns a 201 Created response with a JSON response body containing
the token.